pyproject.toml
[tool.poetry]
name = "gpt_summarizer"
version = "0.1.0"
description = "Summarize text using ChatGPT or a local LLM, with support for multiple large text files, PDF files and translation."
authors = ["Sean Ryan <mr.sean.ryan@gmail.com>"]
readme = "README.md"
license = "MIT"
classifiers = [
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
keywords = ["nlp", "pdf", "summarization", "chatgpt", "pdf-summarizer"]
homepage = "https://github.com/mrseanryan/mrseanryan/gpt-summarizer"
repository = "https://github.com/mrseanryan/mrseanryan/gpt-summarizer"
documentation = "https://github.com/mrseanryan/mrseanryan/gpt-summarizer"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
cornsnake = "^0.0.74"
openai = "1.23.6"
html2text = "2024.2.26"
json5 = "0.9.25"
ollama = "0.2.0"
pydantic = ">=2.7.3,<2.8.0"
pymupdf = "1.24.1"
pyyaml = "6.0.1"

[tool.poetry.group.dev.dependencies]
mypy = "^1.14.0"
ruff = "^0.9.3"

[tool.poetry.group.test.dependencies]
parameterized = "~=0.9.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff.format]
exclude = []

[tool.mypy]
ignore_missing_imports = true
disallow_untyped_defs = true
disallow_any_unimported = true
check_untyped_defs = true
warn_return_any = true
warn_unused_ignores = true
